+++ src/components/StatusBar.jl
@@ -6,7 +6,7 @@
     cc, set_cc = cellcount_signal
     cs, set_cs = connection_signal
 
-    is_dark, set_dark = create_signal(Int32(0))
+    is_dark, set_dark = create_signal(0)
 
     js("if(document.documentElement.classList.contains('dark'))\$1(1)", set_dark)
 
@@ -44,7 +44,7 @@
         Button(:id => "theme-toggle-btn",
             :style => "padding:2px 10px;border-radius:9999px;border:1px solid var(--accent);background:var(--accent-dim);color:var(--accent);font-size:10px;font-family:'JetBrains Mono',monospace;cursor:pointer;transition:all .15s;",
             :on_click => () -> begin
-                set_dark(Int32(1) - is_dark())
+                set_dark(1 - is_dark())
                 js("document.documentElement.classList.toggle('dark')")
                 js("localStorage.setItem('sessions-theme',document.documentElement.classList.contains('dark')?'dark':'light')")
                 js("if(window._sessionsThemeSwitch)_sessionsThemeSwitch()")
